perf: prevent duplicate network requests when fetching opportunities data (#6409)
* perf: use rtk query for zapperAppsBalancesOutput

* perf: use rtk query for zapperUniV2PoolAssetIds

* perf: use react-query for fetchAllOpportunities requests

* chore: actioned code review feedback
