+++ src/pages/Dashboard/Dashboard.tsx
@@ -6,7 +6,7 @@ import { Route, Switch, useRouteMatch } from 'react-router'
 import { Main } from 'components/Layout/Main'
 import { SEO } from 'components/Layout/Seo'
 import { NftTable } from 'components/Nfts/NftTable'
-import { useFetchOpportunities } from 'components/StakingVaults/hooks/useFetchOpporunities'
+import { useFetchOpportunities } from 'components/StakingVaults/hooks/useFetchOpportunities'
 import { RawText } from 'components/Text'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 import { Accounts } from 'pages/Accounts/Accounts'
