+++ src/pages/Defi/views/StakingVaults.tsx
@@ -3,7 +3,7 @@ import { useTranslate } from 'react-polyglot'
 import { Main } from 'components/Layout/Main'
 import { SEO } from 'components/Layout/Seo'
 import { DeFiEarn } from 'components/StakingVaults/DeFiEarn'
-import { useFetchOpportunities } from 'components/StakingVaults/hooks/useFetchOpporunities'
+import { useFetchOpportunities } from 'components/StakingVaults/hooks/useFetchOpportunities'
 
 import { EligibleSlider } from '../components/EligibleSlider'
